+++ apps/vendor-portal/src/components/notification-dropdown/_internal/notification/_internal/DefaultPush.vue
@@ -1,0 +1,24 @@
+<template>
+  <slot name="title" v-bind:title="notification.title"></slot>
+  <vc-hint class="vc-margin-bottom_xs" v-if="notification.description">{{
+    notification.description
+  }}</vc-hint>
+</template>
+
+<script lang="ts">
+import { defineComponent, PropType } from "vue";
+import { PushNotification } from "@virtoshell/api-client";
+
+export default defineComponent({
+  name: "DefaultPush",
+  inheritAttrs: false,
+  props: {
+    notification: {
+      type: Object as PropType<PushNotification>,
+      default: () => ({}),
+    },
+  },
+});
+</script>
+
+<style lang="less" scoped></style>