+++ apps/vendor-portal/src/modules/products/pages/products-edit.vue
@@ -1,6 +1,6 @@
 <template>
   <vc-blade
-    v-loading="loading || !currentCategory"
+    v-loading="loading || productLoading"
     :title="param ? productDetails?.name : $t('PRODUCTS.PAGES.DETAILS.TITLE')"
     width="50%"
     :expanded="expanded"
@@ -74,7 +74,21 @@
                 @update:modelValue="setCategory"
                 :is-disabled="readonly"
                 name="category"
-              ></vc-select>
+              >
+                <template v-slot:item="itemData">
+                  <div
+                    class="vc-flex vc-flex-align_center vc-padding-vertical_s vc-ellipsis"
+                  >
+                    <div class="flex-grow_1 vc-margin-left_l vc-ellipsis">
+                      <div class="vc-ellipsis">{{ itemData.item.path }}</div>
+                      <vc-hint class="vc-ellipsis vc-margin-top_xs">
+                        {{ $t("PRODUCTS.PAGES.DETAILS.FIELDS.CODE") }}:
+                        {{ itemData.item.code }}
+                      </vc-hint>
+                    </div>
+                  </div>
+                </template>
+              </vc-select>
 
               <vc-card
                 :header="$t('PRODUCTS.PAGES.DETAILS.FIELDS.TITLE')"
@@ -178,6 +192,7 @@ import {
   IProperty,
   IPropertyValue,
   PropertyValue,
+  PropertyDictionaryItem,
 } from "@virtoshell/api-client";
 import MpProductStatus from "../components/MpProductStatus.vue";
 import { AssetsDetails } from "@virtoshell/mod-assets";
@@ -237,17 +252,23 @@ export default defineComponent({
     const currentCategory = ref();
     const offersCount = ref(0);
     const categories = ref<ICategory[]>();
+    const productLoading = ref(false);
 
     const reload = async (fullReload: boolean) => {
       if (!modified.value && fullReload) {
-        if (props.param) {
-          await loadProduct({ id: props.param });
-        }
-        categories.value = await fetchCategories();
-        if (productDetails?.categoryId) {
-          currentCategory.value = categories.value?.find(
-            (x) => x.id === productDetails.categoryId
-          );
+        try {
+          productLoading.value = true;
+          if (props.param) {
+            await loadProduct({ id: props.param });
+          }
+          categories.value = await fetchCategories();
+          if (productDetails?.categoryId) {
+            currentCategory.value = categories.value?.find(
+              (x) => x.id === productDetails.categoryId
+            );
+          }
+        } finally {
+          productLoading.value = false;
         }
       }
       //Load offers count to populate widget
@@ -393,7 +414,7 @@ export default defineComponent({
       formData.append("file", files[0]);
       const authToken = await getAccessToken();
       const result = await fetch(
-        `/api/platform/assets?folderUrl=/catalog/${product.value.id}`,
+        `/api/assets?folderUrl=/catalog/${product.value.id}`,
         {
           method: "POST",
           body: formData,
@@ -459,6 +480,14 @@ export default defineComponent({
       });
     };
 
+    async function loadDictionaries(
+      property: IProperty,
+      keyword?: string,
+      skip?: number
+    ) {
+      return await searchDictionaryItems([property.id], keyword, skip);
+    }
+
     let isOffersOpened = false;
 
     return {
@@ -477,6 +506,7 @@ export default defineComponent({
       productDetails,
       readonly: computed(() => props.param && !product.value?.canBeModified),
       statusText,
+      productLoading,
       reload,
       editImages,
       loading: computed(() => loading.value),
@@ -513,7 +543,11 @@ export default defineComponent({
         }
       },
 
-      setPropertyValue(property: IProperty, value: IPropertyValue) {
+      setPropertyValue(
+        property: IProperty,
+        value: IPropertyValue,
+        dictionary?: PropertyDictionaryItem[]
+      ) {
         if (
           typeof value === "object" &&
           Object.prototype.hasOwnProperty.call(value, "length")
@@ -522,28 +556,47 @@ export default defineComponent({
             (item) => new PropertyValue(item)
           );
         } else {
-          if (property.values[0]) {
-            property.values[0].value = value;
-          } else {
+          if (dictionary && dictionary.length) {
+            const valueId = value as string;
+            let valueName;
+            const dictionaryItem = dictionary.find((x) => x.id === valueId);
+            if (dictionaryItem) {
+              valueName = dictionaryItem.alias;
+            } else {
+              valueName = property.name;
+            }
             property.values[0] = new PropertyValue({
-              value,
+              valueId,
+              value: valueName,
               isInherited: false,
             });
+          } else {
+            if (property.values[0]) {
+              property.values[0].value = value;
+            } else {
+              property.values[0] = new PropertyValue({
+                value,
+                isInherited: false,
+              });
+            }
           }
         }
       },
 
-      getPropertyValue(property: IProperty): Record<string, unknown> {
+      getPropertyValue(
+        property: IProperty,
+        isDictionary?: boolean
+      ): Record<string, unknown> {
+        if (isDictionary) {
+          return (
+            property.values[0] &&
+            (property.values[0].valueId as unknown as Record<string, unknown>)
+          );
+        }
         return property.values[0] && property.values[0].value;
       },
 
-      async loadDictionaries(
-        property: IProperty,
-        keyword?: string,
-        skip?: number
-      ) {
-        return await searchDictionaryItems([property.id], keyword, skip);
-      },
+      loadDictionaries,
 
       handleCollapsed(key: string, value: boolean): void {
         localStorage?.setItem(key, `${value}`);